Remove wcutil and replace with new core-element
