--- src/core-element/core-hello/CoreHelloElement.html
@@ -1,4 +1,0 @@
-<label>
-    <span id="hello">Hello</span>
-    <slot id="name"></slot>
-</label>
