--- src/core-element/core-slider/CoreSliderElement.html
@@ -1,5 +1,0 @@
-<!--Slider component-->
-<div id="slider">
-    <span id="slider-bar"></span>
-    <span id="slider-thumb"></span>
-</div>
