+++ src/core-slider/CoreSliderElement.css
