+++ src/core-slider/CoreSliderElement.html
